refactor: consolidate error tests into tests/web/test_errors.py
Merge test_error_show.py, test_trigger_fix.py, and error tests from
test_projects.py into a single test_errors.py mirroring web/errors.py.
